Quick, crisp refactoring
The working hack.lisp demo now resembles the XDG shell basics example
code from Drew Devault's Wayland Book:
<https://wayland-book.com/xdg-shell-basics/example-code.html>

The next hurdle to climb is sending fd's over to the compositor. At this
point, the compositor that I've used for testing (sway on wlroots) reads
the message by wl-shm-create-pool as an error, and hangs up the
connection.

I believe this is due to the differences between Wayflan and Libwayland
message-sending. Libwayland stores both payload data, and file
descriptors, in a ring buffer, to flush it out on a single sendmsg()
syscall. Meanwhile, Wayland's iolib stores only the payload data in
a buffer, and only has the ability to send file descriptors
one-at-a-time through sendmsg().