Replace iolib socket stream with homebrew abstraction
I've been having issues with sending messages that contain file
desctiptors, and I believe this is caused by how iolib sends file
descriptors alongside I/O data: it doesn't. They're sent in separate
messages, in contract to libwayland which buffers both of them and sends
them in the same message.

This commit removes iolib, and replaces it with a homebrew gray-stream
on-top of POSIX sockets, which buffers IO data before sending it off to
sendmsg. As a net bonus, since we're using the sendmsg() syscall, we get
to create an efficient ring buffer through 2-element iovec arrays.

File descriptors is still TODO. This commit just replaces all the
working parts.