Add utility macros, two new examples
This commit takes the utility constants and macros from the Wayland Book
port, and exposes them from the wayflan-client package.

This commit also adds two new examples, global-info.lisp and
cairo-demo.lisp. The global-info demo lists the globals registered for
the Wayland client, and all allowed shm formats. The cairo-demo demo
draws a resizeable window with a rainbow background, text, and an image.
All examples are registered in the new system wayflan/examples.