Replace enum constants with keyword-decoding
This commit replaces defining enums as constants, with decoding enums
into keywords, or a list of keywords for bitfield enums.

For bitfield enum members that takes multiple bits, its keyword is
included in the list so long as all of its bits are set in the integer.
Enum members whose bits contain a subset of the former are also
included.

If a bitfield enum has a member with zero as its value, the integer 0 is
decoded as a singleton list of that member. Otherwise, it is decoded as
an empty list.