Consolidate sockets system into the wire substrate
wayflan/sockets was created as a shim to replace iolib's socket stream,
to send fd's in the same message as mundane I/O. Using gray-streams
creates a small layer of overhead that this commit removes by
consolidating the systems together.

This commit also rewrites the wire and client unit tests, as the wire
API is completely redesigned.