Clarify single-threadedness and server/compositor terminology
This commit clarifies in the README that the library is single-threaded
as it stands right now, but intends to be multi-threadable in the coming
future.

It also clarifies between Wayland servers and compositors, the former
often being casually called the latter since the wl_compositor is the
most critical/visible resource to most users.